🛠[회원가입] : 파일 구조 변경
