+++ app/sign-up/layout.tsx
@@ -1,6 +1,6 @@
 export const metadata = {
   title: "회원가입",
-  description: "Generated by Next.js",
+  description: "토크키트 회원가입",
 };
 
 export default function RootLayout({
